FraudDetApp/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.1.0
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
gradio==4.43.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==3.8
importlib_resources==6.4.5
incremental==24.7.2
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==2.1.1
openai==1.44.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.8
requests==2.32.3
rich==13.8.0
ruff==0.6.4
scikit-learn==1.5.1
scipy==1.14.1
semantic-version==2.10.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.5
threadpoolctl==3.5.0
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
xgboost==2.1.1
zope.interface==7.0.3
